Rename catalog.ts to avoid naming conflict
The GitHub post-commit workflow creates a file `catalog.js`, which
conflicts with the new source file named `catalog.ts`. To avoid this
conflict, `catalog.ts` has been renamed to `catalogData.ts`.
